+++ base_layer/p2p/tests/support/mod.rs
@@ -20,5 +20,36 @@
 //  WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE
 //  USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
-mod ping_pong;
-mod support;
+use rand::{distributions::Alphanumeric, rngs::OsRng, Rng};
+use std::{fmt::Debug, iter, thread, time::Duration};
+
+pub fn random_string(len: usize) -> String {
+    let mut rng = OsRng::new().unwrap();
+    iter::repeat(()).map(|_| rng.sample(Alphanumeric)).take(len).collect()
+}
+
+pub fn assert_change<F, T>(func: F, to: T, poll_count: usize)
+where
+    F: Fn() -> T,
+    T: Eq + Debug,
+{
+    let mut i = 0;
+    loop {
+        let new_val = func();
+        if new_val == to {
+            break;
+        }
+
+        i += 1;
+        if i >= poll_count {
+            panic!(
+                "Value did not change to {:?} within {}ms (last value: {:?})",
+                to,
+                poll_count * 100,
+                new_val,
+            );
+        }
+
+        thread::sleep(Duration::from_millis(100));
+    }
+}
