+++ base_layer/wallet/src/wallet.rs
@@ -75,126 +75,3 @@ impl Wallet {
         })
     }
 }
-
-#[cfg(test)]
-mod test {
-    use super::*;
-    use crate::text_message_service::TextMessageApiResponse;
-    use std::{thread, time::Duration};
-    use tari_comms::{
-        connection::{net_address::NetAddressWithStats, NetAddress, NetAddressesWithStats},
-        control_service::ControlServiceConfig,
-        peer_manager::{peer::PeerFlags, NodeId, Peer},
-        types::{CommsPublicKey, CommsSecretKey},
-    };
-    use tari_crypto::keys::{PublicKey, SecretKey};
-    use tari_p2p::tari_message::{NetMessage, TariMessageType};
-
-    fn create_peer(public_key: CommsPublicKey, net_address: NetAddress) -> Peer {
-        Peer::new(
-            public_key.clone(),
-            NodeId::from_key(&public_key).unwrap(),
-            NetAddressesWithStats::new(vec![NetAddressWithStats::new(net_address.clone())]),
-            PeerFlags::empty(),
-        )
-    }
-
-    #[test]
-    fn test_wallet() {
-        let mut rng = rand::OsRng::new().unwrap();
-
-        let listener_address1: NetAddress = "127.0.0.1:32775".parse().unwrap();
-        let secret_key1 = CommsSecretKey::random(&mut rng);
-        let public_key1 = CommsPublicKey::from_secret_key(&secret_key1);
-        let config1 = WalletConfig {
-            comms: CommsConfig {
-                control_service: ControlServiceConfig {
-                    listener_address: listener_address1.clone(),
-                    socks_proxy_address: None,
-                    accept_message_type: TariMessageType::new(NetMessage::Accept),
-                    requested_outbound_connection_timeout: Duration::from_millis(5000),
-                },
-                socks_proxy_address: None,
-                host: "0.0.0.0".parse().unwrap(),
-                public_key: public_key1.clone(),
-                secret_key: secret_key1,
-            },
-            public_key: public_key1.clone(),
-        };
-        let wallet1 = Wallet::new(config1).unwrap();
-
-        let listener_address2: NetAddress = "127.0.0.1:32776".parse().unwrap();
-        let secret_key2 = CommsSecretKey::random(&mut rng);
-        let public_key2 = CommsPublicKey::from_secret_key(&secret_key2);
-        let config2 = WalletConfig {
-            comms: CommsConfig {
-                control_service: ControlServiceConfig {
-                    listener_address: listener_address2.clone(),
-                    socks_proxy_address: None,
-                    accept_message_type: TariMessageType::new(NetMessage::Accept),
-                    requested_outbound_connection_timeout: Duration::from_millis(5000),
-                },
-                socks_proxy_address: None,
-                host: "0.0.0.0".parse().unwrap(),
-                public_key: public_key2.clone(),
-                secret_key: secret_key2,
-            },
-            public_key: public_key2.clone(),
-        };
-
-        let wallet2 = Wallet::new(config2).unwrap();
-
-        wallet1
-            .comms_services
-            .peer_manager
-            .add_peer(create_peer(public_key2.clone(), listener_address2))
-            .unwrap();
-
-        wallet2
-            .comms_services
-            .peer_manager
-            .add_peer(create_peer(public_key1.clone(), listener_address1))
-            .unwrap();
-
-        wallet1
-            .text_message_service
-            .send_text_message(public_key2.clone(), "Say Hello,".to_string())
-            .unwrap();
-
-        wallet2
-            .text_message_service
-            .send_text_message(public_key1.clone(), "hello?".to_string())
-            .unwrap();
-
-        wallet1
-            .text_message_service
-            .send_text_message(public_key2.clone(), "to my little friend!".to_string())
-            .unwrap();
-
-        thread::sleep(Duration::from_millis(500));
-
-        if let TextMessageApiResponse::TextMessages(msgs) = wallet1.text_message_service.get_text_messages().unwrap() {
-            assert_eq!(msgs.sent_messages.len(), 2);
-            assert_eq!(msgs.received_messages.len(), 1);
-        } else {
-            assert!(false);
-        }
-
-        if let TextMessageApiResponse::TextMessages(msgs) = wallet2.text_message_service.get_text_messages().unwrap() {
-            assert_eq!(msgs.received_messages.len(), 2);
-            assert_eq!(msgs.sent_messages.len(), 1);
-        } else {
-            assert!(false);
-        }
-
-        wallet1.ping_pong_service.ping(public_key2.clone()).unwrap();
-        wallet2.ping_pong_service.ping(public_key1.clone()).unwrap();
-
-        thread::sleep(Duration::from_millis(500));
-
-        assert_eq!(wallet1.ping_pong_service.ping_count().unwrap(), 1);
-        assert_eq!(wallet1.ping_pong_service.pong_count().unwrap(), 1);
-        assert_eq!(wallet2.ping_pong_service.ping_count().unwrap(), 1);
-        assert_eq!(wallet2.ping_pong_service.pong_count().unwrap(), 1);
-    }
-}
