+++ base_layer/wallet/tests/support/mod.rs
@@ -21,3 +21,4 @@
 // USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
 pub mod comms_and_services;
+pub mod utils;
